feat - make socket modular
