+++ src/app/chat/layout.tsx
@@ -1,20 +1,20 @@
 import type { Metadata } from "next";
 
-import ChatContactLayout from "@/components/layout/chat-layout";
+import ChatContactLayout from "@/app/chat/components/chat-layout";
 
 export const metadata: Metadata = {
-	title: "Chat App - Room",
-	description: "Chat with your friends!",
+  title: "Chat App - Room",
+  description: "Chat with your friends!",
 };
 
 export default function ChatLayout({
-	children,
+  children,
 }: Readonly<{
-	children: React.ReactNode;
+  children: React.ReactNode;
 }>) {
-    return <main>
-        <ChatContactLayout>
-            {children}
-        </ChatContactLayout>
+  return (
+    <main>
+      <ChatContactLayout>{children}</ChatContactLayout>
     </main>
+  );
 }
