+++ src/stores/socket-store.ts
@@ -17,6 +17,7 @@ export type SocketActions = {
   setSocket: (newSocket: Socket) => void;
   joinSocketRoom: (roomID: string) => void;
   confirmJoinRoom: (roomID: string) => void;
+  sendMessageToRoom: (roomID: string, chatMessage: string) => void;
 };
 
 export type SocketStore = SocketState & SocketActions;
@@ -96,6 +97,7 @@ export const createSocketStore = (
             });
           }
           router.push(`/chat/${roomID}`);
+          socket.emit("joined-room", roomID);
           toast({
             description: "Joined room successfully!",
           });
@@ -104,6 +106,19 @@ export const createSocketStore = (
 
       confirmJoinRoom: (roomID: string) => {
         socket.emit("joined-room", roomID);
+      },
+
+      sendMessageToRoom: (roomID: string, chatMessage: string) => {
+        if (chatMessage.trim().length === 0) {
+          return;
+        }
+        if (!socket) {
+          return toast({
+            variant: "destructive",
+            title: "Socket connection doesn't exists!",
+          });
+        }
+        socket.emit("send-message", roomID, chatMessage);
       },
     };
   });
